Time: 18 ms (28.12%), Space: 17.3 MB (48.00%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    int widthOfBinaryTree(TreeNode* root) {
+    long long widthOfBinaryTree(TreeNode* root) {
         if(!root) return 0;
         if(!root->left and !root->right) return 1;
